+++ editor/AusgangErweitert.java
@@ -24,8 +24,8 @@ public class AusgangErweitert implements Serializable {
 	Point[] points;
 	boolean von1nach2 = true;
 	boolean von2nach1 = true;
-	int bez_von1nach2 = Ausgang.EIGENE;
-	int bez_von2nach1 = Ausgang.EIGENE;
+	byte bez_von1nach2 = Ausgang.EIGENE;
+	byte bez_von2nach1 = Ausgang.EIGENE;
 	String eigeneBez_von1nach2 = "";
 	String eigeneBez_von2nach1 = "";
